requirements.txt
mysql-connector-python==9.5.0
pandas==2.2.3
streamlit==1.51.0
plotly==5.24.1
SQLAlchemy==2.0.44
PyMySQL==1.1.2
python-dateutil==2.9.0
